requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --strip-extras requirements.in
#
annotated-types==0.7.0
    # via pydantic
appdirs==1.4.4
    # via pyppeteer
audioop-lts==0.2.2
    # via
    #   speechrecognition
    #   standard-aifc
beartype==0.21.0
    # via plum-dispatch
beautifulsoup4==4.13.4
    # via bs4
blis==1.3.0
    # via thinc
bs4==0.0.2
    # via requests-html
catalogue==2.0.10
    # via
    #   spacy
    #   srsly
    #   thinc
certifi==2025.8.3
    # via
    #   pyppeteer
    #   requests
charset-normalizer==3.4.3
    # via requests
click==8.2.1
    # via typer
cloudpathlib==0.21.1
    # via weasel
cloudpickle==3.1.1
    # via fastai
colorama==0.4.6
    # via
    #   click
    #   tqdm
    #   wasabi
comtypes==1.4.11
    # via pyttsx3
confection==0.1.5
    # via
    #   thinc
    #   weasel
contourpy==1.3.3
    # via matplotlib
cssselect==1.3.0
    # via pyquery
cycler==0.12.1
    # via matplotlib
cymem==2.0.11
    # via
    #   preshed
    #   spacy
    #   thinc
fake-useragent==2.2.0
    # via requests-html
fastai==2.8.3
    # via -r requirements.in
fastcore==1.8.7
    # via
    #   fastai
    #   fastdownload
    #   fasttransform
fastdownload==0.0.7
    # via fastai
fastprogress==1.0.3
    # via
    #   fastai
    #   fastdownload
fasttransform==0.0.2
    # via fastai
filelock==3.18.0
    # via
    #   huggingface-hub
    #   torch
fonttools==4.59.0
    # via matplotlib
fsspec==2025.3.0
    # via
    #   huggingface-hub
    #   torch
huggingface-hub==0.34.4
    # via -r requirements.in
idna==3.10
    # via requests
importlib-metadata==8.7.0
    # via pyppeteer
jinja2==3.1.6
    # via
    #   spacy
    #   torch
joblib==1.5.1
    # via scikit-learn
kiwisolver==1.4.9
    # via matplotlib
langcodes==3.5.0
    # via spacy
language-data==1.3.0
    # via langcodes
lxml==6.0.0
    # via pyquery
marisa-trie==1.2.1
    # via language-data
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
matplotlib==3.10.5
    # via fastai
mdurl==0.1.2
    # via markdown-it-py
mpmath==1.3.0
    # via sympy
murmurhash==1.0.13
    # via
    #   preshed
    #   spacy
    #   thinc
networkx==3.5
    # via torch
numpy==2.3.2
    # via
    #   blis
    #   contourpy
    #   matplotlib
    #   pandas
    #   scikit-learn
    #   scipy
    #   spacy
    #   thinc
    #   torchvision
packaging==25.0
    # via
    #   fastai
    #   fastcore
    #   huggingface-hub
    #   matplotlib
    #   spacy
    #   thinc
    #   weasel
pandas==2.3.1
    # via fastai
parse==1.20.2
    # via requests-html
pillow==11.3.0
    # via
    #   -r requirements.in
    #   fastai
    #   matplotlib
    #   torchvision
plum-dispatch==2.5.7
    # via
    #   fastai
    #   fasttransform
preshed==3.0.10
    # via
    #   spacy
    #   thinc
pydantic==2.11.7
    # via
    #   confection
    #   spacy
    #   thinc
    #   weasel
pydantic-core==2.33.2
    # via pydantic
pyee==11.1.1
    # via pyppeteer
pygments==2.19.2
    # via rich
pyparsing==3.2.3
    # via matplotlib
pypiwin32==223
    # via pyttsx3
pyppeteer==2.0.0
    # via requests-html
pyquery==2.0.1
    # via requests-html
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
python-dotenv==1.1.1
    # via -r requirements.in
pyttsx3==2.99
    # via -r requirements.in
pytz==2025.2
    # via pandas
pywin32==311
    # via
    #   pypiwin32
    #   pyttsx3
pyyaml==6.0.2
    # via
    #   fastai
    #   huggingface-hub
readchar==4.2.1
    # via -r requirements.in
requests==2.32.4
    # via
    #   fastai
    #   huggingface-hub
    #   requests-html
    #   spacy
    #   weasel
requests-html==0.10.0
    # via -r requirements.in
rich==14.1.0
    # via
    #   plum-dispatch
    #   typer
scikit-learn==1.7.1
    # via fastai
scipy==1.16.1
    # via
    #   fastai
    #   scikit-learn
shellingham==1.5.4
    # via typer
six==1.17.0
    # via python-dateutil
smart-open==7.3.0.post1
    # via weasel
soupsieve==2.7
    # via beautifulsoup4
spacy==3.8.7
    # via fastai
spacy-legacy==3.0.12
    # via spacy
spacy-loggers==1.0.5
    # via spacy
speechrecognition==3.14.3
    # via -r requirements.in
srsly==2.5.1
    # via
    #   confection
    #   spacy
    #   thinc
    #   weasel
standard-aifc==3.13.0
    # via speechrecognition
standard-chunk==3.13.0
    # via standard-aifc
sympy==1.14.0
    # via torch
thinc==8.3.6
    # via spacy
threadpoolctl==3.6.0
    # via scikit-learn
torch==2.8.0
    # via
    #   -r requirements.in
    #   fastai
    #   torchvision
torchvision==0.23.0
    # via fastai
tqdm==4.67.1
    # via
    #   huggingface-hub
    #   pyppeteer
    #   spacy
typer==0.16.0
    # via
    #   spacy
    #   weasel
typing-extensions==4.14.1
    # via
    #   beautifulsoup4
    #   huggingface-hub
    #   plum-dispatch
    #   pydantic
    #   pydantic-core
    #   pyee
    #   speechrecognition
    #   torch
    #   typer
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
tzdata==2025.2
    # via pandas
urllib3==1.26.20
    # via
    #   pyppeteer
    #   requests
w3lib==2.3.1
    # via requests-html
wasabi==1.1.3
    # via
    #   spacy
    #   thinc
    #   weasel
weasel==0.4.1
    # via spacy
websockets==10.4
    # via pyppeteer
wrapt==1.17.3
    # via smart-open
zipp==3.23.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
